Rename dataformat -> raggedarray
